Move traverseFilesSync to utils and ensure strict mode everywhere
