+++ src/postman-replacements.js
@@ -1,3 +1,4 @@
+'use strict';
 const utils = require('./utils');
 
 const replaceHeaders = (postmanCollection, headerReplacements) => {
